Updates
Some corrections..
Troubles:
After new changes, there is still problem. If i send a string with "OK" in it, i receive same string back. If i send some strings which has no "OK" in it and after that if i send a string with "OK" in it again, then i don't receive any string at all. I tried to solve it, but i couldn't. Could you check it?